backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
flake8==7.3.0
gunicorn==25.2.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
setuptools==82.0.1
